Issue #99: Refactored `CompareCore` as es6 module.
