+++ bw_matchbox/assets/css/compare.css
@@ -5,7 +5,7 @@
   --compare-selected-bg-color3: rgb(30 174 219 / 30%);
   --compare-selected-bg-color4: rgb(30 174 219 / 40%);
   --compare-selected-bg-color5: rgb(30 174 219 / 50%);
-  --compare-selected-animation-time: 3s;
+  --compare-selected-animation-time: 1.5s;
 }
 
 /* Collapsed tools (shown only if collapsed rows are extist */
